requirements.txt
gspread>=6.0
streamlit>=1.30
plotly>=5.18
pyyaml>=6.0
pytest>=8.0
